Auto increment feature restored (changed data type from long to int)

+++ src/main/java/eu/codeacademy/project/questions/repository/QuestionsRepository.java
@@ -3,5 +3,5 @@ package eu.codeacademy.project.questions.repository;
 import eu.codeacademy.project.questions.entity.Questions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionsRepository extends JpaRepository<Questions, Long> {
+public interface QuestionsRepository extends JpaRepository<Questions, Integer> {
 }
